1 print list

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -3,13 +3,13 @@
 /**
  * list_len - function that returns the number
  * of elements in a linked list_t list.
- * @h: the list
+ * @h:the list
  *
- * Return: The element
+ * Return:The element
  */
 size_t list_len(const list_t *h)
 {
-	int i = 0;
+	size_t i  = 0;
 
 	while (h != NULL)
 	{
